requirements.txt
# LIME — Phase 1 Dependencies

# API Server
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==13.1
python-multipart==0.0.12

# Audio Capture
sounddevice==0.5.1
numpy==1.26.4
scipy==1.13.1

# Voice Activity Detection
silero-vad==5.1.2

# Transcription (local)
faster-whisper==1.1.0

# Transcription (cloud)
deepgram-sdk==3.8.0
assemblyai==0.33.0

# Speaker Diarization
pyannote.audio==3.3.2
torch==2.5.1
torchaudio==2.5.1

# Audio Compression
ffmpeg-python==0.2.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Configuration
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# CLI
typer==0.14.0
rich==13.9.4

# Utilities
aiofiles==24.1.0
httpx==0.28.1
